Rename page checksum error to file error in info text output.
"file error" makes it clear that other errors may be reported. If checksum-page is true in the manifest but no checksum-page-error list is provided then the error is in alignment, i.e. the file size is not a multiple of the page size, with allowances made for a valid-looking partial page at the end of the file.