fix: remove trailing slashes from repo-ls paths
Trailing slashes in at least some of the backend repo types were preventing `repo-ls` from
displaying any content (presumably backend-specific behavior).  Since the path with the slash should
be equivalent to the path without the slash, just remove it if provided by the user to prevent this
little UI wart.