Disconnect help.auto.c from build-help in Makefile
When there is an issue with the system library path during building, the
build-help rule will fail during executing ./build-help with the effect
that main.c wont build. Break out help.auto.c generation from the build-
help stage to allow it to be re-executed when the library path has been
corrected.